Rename `errorParser` to `utils`, move `findOrCreateUser` to `utils`
