+++ config/passport.js
@@ -4,75 +4,59 @@ const GitHubStrategy = require('passport-github2').Strategy;
 const GoogleStrategy = require('passport-google-oauth20').Strategy;
 
 const User = require('../models/user');
+const { findOrCreateUser } = require('../utils/utils');
 
 const ExtractJwt = passportJWT.ExtractJwt;
 const JWTStrategy = passportJWT.Strategy;
 
 const options = {
-    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-    secretOrKey: process.env.JWT_SECRET
+  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+  secretOrKey: process.env.JWT_SECRET
 };
 
 const JwtStrategy = new JWTStrategy(options, (payload, next) => {
-    console.log('payload received', payload);
-    const id = payload.id;
-    User.findOne({ _id: id }).then(user => {
-        console.log('user', user);
-        if (user) {
-            next(null, user);
-        } else {
-            next(null, false);
-        }
-    });
+  console.log('payload received', payload);
+  const id = payload.id;
+  User.findOne({ _id: id }).then(user => {
+    console.log('user', user);
+    if (user) {
+      next(null, user);
+    } else {
+      next(null, false);
+    }
+  });
 });
 
 passport.use(
-    new GitHubStrategy({
-        clientID: process.env.GITHUB_CLIENT_ID,
-        clientSecret: process.env.GITHUB_CLIENT_SECRET,
-        callbackURL: '/auth/github/redirect'
-    }, (accessToken, refreshToken, profile, done) => {
-        process.nextTick(() => {
-            findOrCreateUser(profile, done);
-        });
-    })
+  new GitHubStrategy(
+    {
+      clientID: process.env.GITHUB_CLIENT_ID,
+      clientSecret: process.env.GITHUB_CLIENT_SECRET,
+      callbackURL: '/auth/github/redirect'
+    },
+    (accessToken, refreshToken, profile, done) => {
+      process.nextTick(() => {
+        findOrCreateUser(profile, done);
+      });
+    }
+  )
 );
 
 passport.use(
-    new GoogleStrategy({
-        clientID: process.env.GOOGLE_CLIENT_ID,
-        clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-        callbackURL: '/auth/google/redirect'
-    }, (accessToken, refreshToken, profile, done) => {
-        process.nextTick(() => {
-            findOrCreateUser(profile, done);
-        });
-    })
+  new GoogleStrategy(
+    {
+      clientID: process.env.GOOGLE_CLIENT_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+      callbackURL: '/auth/google/redirect'
+    },
+    (accessToken, refreshToken, profile, done) => {
+      process.nextTick(() => {
+        findOrCreateUser(profile, done);
+      });
+    }
+  )
 );
 
-const findOrCreateUser = (profile, done) => {
-    User.findOne({ email: profile.emails[0].value })
-        .then(user => {
-            if (user) {
-                console.log('user found');
-                done(null, user);
-            } else {
-                new User({
-                    username: profile.displayName,
-                    email: profile.emails[0].value,
-                    thumbnail: profile.photos[0].value
-                })
-                    .generateAuthToken()
-                    .then(newUser => {
-                        console.log('created new user', newUser);
-                        done(null, newUser);
-                    });
-            }
-        })
-        .catch(err => done(err));
-};
-
-
 module.exports = {
-    JwtStrategy
+  JwtStrategy
 };
